pyproject.toml
[project]
name = "megahack-2026-merge-conflictx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.20",
    "numpy>=2.4.3",
    "pdfplumber>=0.11.9",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=3.0.1",
    "faiss-cpu>=1.8.0",
    "soundfile>=0.13.1",
    "pydantic>=2.12.5",
]
